[executor] delete transaction_block.rs
