+++ execution/executor/src/lib.rs
@@ -4,19 +4,18 @@
 
 mod block_processor;
 mod block_tree;
-pub mod transaction_block;
 
 #[cfg(test)]
 mod executor_test;
 #[cfg(test)]
 mod mock_vm;
 
-use crate::{block_processor::BlockProcessor, transaction_block::ProcessedVMOutput};
+use crate::block_processor::BlockProcessor;
 use config::config::NodeConfig;
 use crypto::{
     hash::{
-        TransactionAccumulatorHasher, ACCUMULATOR_PLACEHOLDER_HASH, GENESIS_BLOCK_ID,
-        PRE_GENESIS_BLOCK_ID, SPARSE_MERKLE_PLACEHOLDER_HASH,
+        EventAccumulatorHasher, TransactionAccumulatorHasher, ACCUMULATOR_PLACEHOLDER_HASH,
+        GENESIS_BLOCK_ID, PRE_GENESIS_BLOCK_ID, SPARSE_MERKLE_PLACEHOLDER_HASH,
     },
     HashValue,
 };
@@ -24,6 +23,9 @@ use failure::{format_err, Result};
 use futures::{channel::oneshot, executor::block_on};
 use lazy_static::lazy_static;
 use libra_types::{
+    account_address::AccountAddress,
+    account_state_blob::AccountStateBlob,
+    contract_event::ContractEvent,
     crypto_proxies::LedgerInfoWithSignatures,
     ledger_info::LedgerInfo,
     proof::accumulator::InMemoryAccumulator,
@@ -34,6 +36,7 @@ use logger::prelude::*;
 use scratchpad::SparseMerkleTree;
 use serde::{Deserialize, Serialize};
 use std::collections::BTreeMap;
+use std::collections::HashMap;
 use std::{
     marker::PhantomData,
     sync::{mpsc, Arc, Mutex},
@@ -102,6 +105,119 @@ impl ExecutedState {
             version: 0,
             validators: None,
         }
+    }
+}
+
+/// The entire set of data associated with a transaction. In addition to the output generated by VM
+/// which includes the write set and events, this also has the in-memory trees.
+#[derive(Clone, Debug)]
+pub struct TransactionData {
+    /// Each entry in this map represents the new blob value of an account touched by this
+    /// transaction. The blob is obtained by deserializing the previous blob into a BTreeMap,
+    /// applying relevant portion of write set on the map and serializing the updated map into a
+    /// new blob.
+    account_blobs: HashMap<AccountAddress, AccountStateBlob>,
+
+    /// The list of events emitted during this transaction.
+    events: Vec<ContractEvent>,
+
+    /// The execution status set by the VM.
+    status: TransactionStatus,
+
+    /// The in-memory Sparse Merkle Tree after the write set is applied. This is `Rc` because the
+    /// tree has uncommitted state and sometimes `StateVersionView` needs to have a pointer to the
+    /// tree so VM can read it.
+    state_tree: Arc<SparseMerkleTree>,
+
+    /// The in-memory Merkle Accumulator that has all events emitted by this transaction.
+    event_tree: Arc<InMemoryAccumulator<EventAccumulatorHasher>>,
+
+    /// The amount of gas used.
+    gas_used: u64,
+
+    /// The number of newly created accounts.
+    num_account_created: usize,
+}
+
+impl TransactionData {
+    fn new(
+        account_blobs: HashMap<AccountAddress, AccountStateBlob>,
+        events: Vec<ContractEvent>,
+        status: TransactionStatus,
+        state_tree: Arc<SparseMerkleTree>,
+        event_tree: Arc<InMemoryAccumulator<EventAccumulatorHasher>>,
+        gas_used: u64,
+        num_account_created: usize,
+    ) -> Self {
+        TransactionData {
+            account_blobs,
+            events,
+            status,
+            state_tree,
+            event_tree,
+            gas_used,
+            num_account_created,
+        }
+    }
+
+    fn account_blobs(&self) -> &HashMap<AccountAddress, AccountStateBlob> {
+        &self.account_blobs
+    }
+
+    fn events(&self) -> &[ContractEvent] {
+        &self.events
+    }
+
+    fn status(&self) -> &TransactionStatus {
+        &self.status
+    }
+
+    fn state_root_hash(&self) -> HashValue {
+        self.state_tree.root_hash()
+    }
+
+    fn event_root_hash(&self) -> HashValue {
+        self.event_tree.root_hash()
+    }
+
+    fn gas_used(&self) -> u64 {
+        self.gas_used
+    }
+
+    fn num_account_created(&self) -> usize {
+        self.num_account_created
+    }
+
+    fn prune_state_tree(&self) {
+        self.state_tree.prune()
+    }
+}
+
+/// Generated by processing VM's output.
+#[derive(Debug, Clone)]
+pub struct ProcessedVMOutput {
+    /// The entire set of data associated with each transaction.
+    transaction_data: Vec<TransactionData>,
+
+    /// The in-memory Merkle Accumulator and state Sparse Merkle Tree after appending all the
+    /// transactions in this set.
+    executed_trees: ExecutedTrees,
+}
+
+impl ProcessedVMOutput {
+    pub fn new(transaction_data: Vec<TransactionData>, executed_trees: ExecutedTrees) -> Self {
+        ProcessedVMOutput {
+            transaction_data,
+            executed_trees,
+        }
+    }
+
+    pub fn transaction_data(&self) -> &[TransactionData] {
+        &self.transaction_data
+    }
+
+    pub fn executed_trees(&self) -> &ExecutedTrees {
+        &self.executed_trees
     }
 }
